Refactor: Label에서 ApplicationStatusBadge로 컴포넌트 이름을 변경한다
