+++ @types/custom-types/emotion.d.ts
@@ -8,6 +8,6 @@ declare module '@emotion/react' {
     input: import('@/styles').InputTheme;
     a11y: import('@/styles').A11yTheme;
     navigation: import('@/styles').NavigationTheme;
-    label: import('@/styles').LabelTheme;
+    badge: import('@/styles').BadgeTheme;
   }
 }
