+++ src/components/common/index.ts
@@ -7,4 +7,4 @@ export { default as Checkbox } from './Checkbox/Checkbox.component';
 export { default as ModalWrapper } from './ModalWrapper/ModalWrapper.component';
 export { default as RadioButton } from './RadioButton/RadioButton.component';
 export { default as ToggleButton } from './ToggleButton/ToggleButton.component';
-export { default as Label } from './Label/Label.component';
+export { default as Label } from './ApplicationStatusBadge/ApplicationStatusBadge.component';
