+++ src/styles/themes/index.ts
@@ -5,4 +5,4 @@ export * from './button';
 export * from './theme';
 export * from './input';
 export * from './navigation';
-export * from './label';
+export * from './badge';
