+++ src/styles/themes/theme.ts
@@ -4,7 +4,7 @@ import { fonts } from './fonts';
 import { button } from './button';
 import { input } from './input';
 import { navigation } from './navigation';
-import { label } from './label';
+import { badge } from './badge';
 
 export const theme = {
   a11y,
@@ -13,5 +13,5 @@ export const theme = {
   button,
   input,
   navigation,
-  label,
+  badge,
 };
